added mapper from meteringPointType and settlementMethoad to timeSeriesType
