+++ source/dotnet/wholesale-api/Edi/Mappers/TimeSeriesTypeMapper.cs
@@ -12,50 +12,46 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.EDI.Exceptions;
 using Energinet.DataHub.Wholesale.EDI.Models;
-using TimeSeriesTypeContract = Energinet.DataHub.Edi.Responses.TimeSeriesType;
 
 namespace Energinet.DataHub.Wholesale.EDI.Mappers;
 
-using CalculationTimeSeriesType = Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.EnergyResults.TimeSeriesType;
-
 public static class TimeSeriesTypeMapper
 {
-    public static CalculationTimeSeriesType MapTimeSeriesTypeFromEdi(TimeSeriesType timeSeriesType)
+    public static TimeSeriesType MapTimeSeriesType(Energinet.DataHub.Edi.Requests.TimeSeriesType timeSeriesType, string meteringPointType, string? settlementMethod)
     {
-        return timeSeriesType switch {
-            TimeSeriesType.Production => CalculationTimeSeriesType.Production,
-            TimeSeriesType.FlexConsumption => CalculationTimeSeriesType.FlexConsumption,
-            TimeSeriesType.TotalConsumption => CalculationTimeSeriesType.TotalConsumption,
-            TimeSeriesType.NetExchangePerGa => CalculationTimeSeriesType.NetExchangePerGa,
-            TimeSeriesType.NonProfiledConsumption => CalculationTimeSeriesType.NonProfiledConsumption,
-            _ => throw new InvalidOperationException($"Unknown time series type: {timeSeriesType}"),
-        };
-    }
-
-    public static TimeSeriesTypeContract MapTimeSeriesTypeFromCalculationsResult(CalculationTimeSeriesType timeSeriesType)
-    {
-        return timeSeriesType switch
+        // TODO: Delete this, when EDI has updated their contract
+        if (string.IsNullOrWhiteSpace(meteringPointType))
         {
-            CalculationTimeSeriesType.Production => TimeSeriesTypeContract.Production,
-            CalculationTimeSeriesType.NonProfiledConsumption => TimeSeriesTypeContract.NonProfiledConsumption,
-            CalculationTimeSeriesType.TotalConsumption => TimeSeriesTypeContract.TotalConsumption,
-            CalculationTimeSeriesType.FlexConsumption => TimeSeriesTypeContract.FlexConsumption,
-            CalculationTimeSeriesType.NetExchangePerGa => TimeSeriesTypeContract.NetExchangePerGa,
-            CalculationTimeSeriesType.GridLoss => throw new NotSupportedTimeSeriesTypeException(
-                $"{timeSeriesType} is not a supported TimeSeriesType For AggregatedTimeSeriesRequestAccepted response."),
-            CalculationTimeSeriesType.TempProduction => throw new NotSupportedTimeSeriesTypeException(
-                $"{timeSeriesType} is not a supported TimeSeriesType For AggregatedTimeSeriesRequestAccepted response."),
-            CalculationTimeSeriesType.NegativeGridLoss => throw new NotSupportedTimeSeriesTypeException(
-                $"{timeSeriesType} is not a supported TimeSeriesType For AggregatedTimeSeriesRequestAccepted response."),
-            CalculationTimeSeriesType.PositiveGridLoss => throw new NotSupportedTimeSeriesTypeException(
-                $"{timeSeriesType} is not a supported TimeSeriesType For AggregatedTimeSeriesRequestAccepted response."),
-            CalculationTimeSeriesType.TempFlexConsumption => throw new NotSupportedTimeSeriesTypeException(
-                $"{timeSeriesType} is not a supported TimeSeriesType For AggregatedTimeSeriesRequestAccepted response."),
-            CalculationTimeSeriesType.NetExchangePerNeighboringGa => throw new NotSupportedTimeSeriesTypeException(
-                $"{timeSeriesType} is not a supported TimeSeriesType For AggregatedTimeSeriesRequestAccepted response."),
-            _ => throw new ArgumentOutOfRangeException($"Unknown time series type {nameof(timeSeriesType)}"),
+            return timeSeriesType switch
+            {
+                Edi.Requests.TimeSeriesType.Production => TimeSeriesType.Production,
+                Edi.Requests.TimeSeriesType.FlexConsumption => TimeSeriesType.FlexConsumption,
+                Edi.Requests.TimeSeriesType.NonProfiledConsumption => TimeSeriesType
+                    .NonProfiledConsumption,
+                Edi.Requests.TimeSeriesType.NetExchangePerGa => TimeSeriesType.NetExchangePerGa,
+                Edi.Requests.TimeSeriesType.NetExchangePerNeighboringGa => TimeSeriesType
+                    .NetExchangePerNeighboringGa,
+                Edi.Requests.TimeSeriesType.TotalConsumption => TimeSeriesType.TotalConsumption,
+                Edi.Requests.TimeSeriesType.Unspecified => throw new InvalidOperationException(
+                    "Unknown time series type"),
+                _ => throw new InvalidOperationException("Unknown time series type"),
+            };
+        }
+
+        return meteringPointType switch
+        {
+            "E18" => TimeSeriesType.Production,
+            "E20" => TimeSeriesType.NetExchangePerGa,
+            "E17" => settlementMethod switch
+            {
+                "E02" => TimeSeriesType.NonProfiledConsumption,
+                "D01" => TimeSeriesType.FlexConsumption,
+                var method when
+                    string.IsNullOrWhiteSpace(method) => TimeSeriesType.TotalConsumption,
+                _ => throw new InvalidOperationException("Unknown time series type"),
+            },
+            _ => throw new InvalidOperationException("Unknown time series type"),
         };
     }
 }
